Moved into rc6 package and created setup.py
