+++ rc6/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["ops.encrypt", "ops.decrypt", "helpers"]
+
+from ops import encrypt, decrypt
+from helpers import *
